refactor: move offer/notification views out of legacy adapters
Relocate SessionOfferClaimView, SessionOfferDeclineView and
NotificationsMarkReadView to gates/web/django/chronology/offers.py.
They were already fully on request.services, so this is a pure move
with no behavior change; URL names stay in the web: namespace.

Drops legacy-loc by 64.
