+++ src/ludamus/adapters/web/django/urls.py
@@ -1,6 +1,7 @@
 from django.urls import URLPattern, URLResolver, include, path
 from django.views.generic.base import RedirectView, TemplateView
 
+from ludamus.gates.web.django.chronology import offers
 from ludamus.gates.web.django.chronology.urls import urlpatterns as chronology_gate_urls
 from ludamus.gates.web.django.crowd.urls import urlpatterns as crowd_gate_urls
 from ludamus.gates.web.django.notice_board.urls import (
@@ -34,12 +35,12 @@ chronology_urls = [
     ),
     path(
         "offer/<str:token>/claim/",
-        views.SessionOfferClaimView.as_view(),
+        offers.SessionOfferClaimView.as_view(),
         name="offer-claim",
     ),
     path(
         "offer/<str:token>/decline/",
-        views.SessionOfferDeclineView.as_view(),
+        offers.SessionOfferDeclineView.as_view(),
         name="offer-decline",
     ),
 ]
@@ -49,7 +50,7 @@ urlpatterns = [
     path("events/", views.EventsPageView.as_view(), name="events"),
     path(
         "notifications/do/mark-read",
-        views.NotificationsMarkReadView.as_view(),
+        offers.NotificationsMarkReadView.as_view(),
         name="notifications-mark-read",
     ),
     path("design/", views.DesignPageView.as_view(), name="design"),
